Add template implementations of `*Manager` classes
These correspond to the specifications in `DESIGN.md`. They are likely
to be used internally by the `Item` and/or `Crop` objects so that a
massive rewrite can be avoided.

Having Manager classes would help to manage the changes introduced by
events (as specified in `DESIGN.md`) because the information that they
store only resides in one place.