Make crop harvesting handle trees in `Plot.harvest`
This lets trees be harvested multiple times during their lifetime.
Then after some time, the tree "dies" and is no longer planted in the
plot.

It enables later harvests by planting the tree again (calling
`Plot.plant`) but ensuring it is not as if the tree is being planted for
the *first time*. This is an important detail because the time upon
first planting is used in the calculation for scheduled time to die.

This may be a temporary solution which will stay in place until a cleaner,
more modular approach is made.

Also fixes this line in `Plot.harvest` from:

```python
lifetime = random.randint(self.crop.min_lifetime, self.crop.max_lifetime)```

to:

```python
lifetime = random.randint(self.crop.min_lifetime, self.crop.max_lifetime) * 60```

TODO: Make the `time` field of `Crop`s only be in seconds. This makes
the code simpler because time is already handled and given in seconds by
the stdlib.