Add help descriptions for commands
This adds docstrings to the command coroutines, which gets used by
`discord.py` to generate the brief and long-form help descriptions for
it.

The names of their parameters should be changed so that the usage is a
lot easier to understand. This may mean having to change their semantics
a bit, e.g., `dgive`, `buy`, `sell` MUST have its first argument as an
`int` OR the first argument is an `Optional[int]`.