Fix help message formatting for empty help command
It previously had a 4 space indent, which had every command starting
from the second command to become indented by 4 spaces. An easier fix
would probably have been to initialise `text` with 4 spaces, but this is
a more desirable help command output.

Furthermore, Discord strips messages of leading whitespaces, so if all
the commands were indented, then none of them would *appear* to be.