Fix default value bug in `Container` constructor
Because the default value was a list, it means that *all* calls to
`Container.__init__` have a default value of that *same* list object.
This means that any changes to that list get propagated to *every*
`Container` instance that was instantiated with the default value.

This is because lists are mutable, of course.